Changes for beta release
Summary:
Added beta

Added possibility to perform UPSERT

Added new option - `upsert`.
Default value is `false`.
This diff is based on https://grizzly.internal.memcompute.com/D61939
**Design doc/spec**:
**Docs impact**: none
**Preliminary Reviewer(s)**:
**Final Reviewer**:

Test Plan: https://app.circleci.com/pipelines/github/memsql/singlestore-kafka-connector/178/workflows/da915645-fd79-4dc1-a002-81c3234d3b66

Reviewers: pmishchenko-ua, mshcherbina-ua

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D61991

+++ src/test/java/com/singlestore/kafka/integration/IntegrationBase.java
@@ -56,9 +56,9 @@ public class IntegrationBase {
     }
 
     public static void put(Map<String, String> props, List<SinkRecord> records) throws SQLException {
-        put(props, records, null);
+        put(props, records, null, true);
     }
-    public static void put(Map<String, String> props, List<SinkRecord> records, String query) throws SQLException {
+    public static void put(Map<String, String> props, List<SinkRecord> records, String query, boolean dropTableBeforeWrite) throws SQLException {
         props.put(SingleStoreSinkConfig.DDL_ENDPOINT, "localhost:5506");
         props.put(SingleStoreSinkConfig.CONNECTION_DATABASE, "testdb");
         String password;
@@ -67,7 +67,9 @@ public class IntegrationBase {
         }
         props.put(SingleStoreSinkConfig.METADATA_TABLE_ALLOW, "false");
 
-        executeQuery(String.format("DROP TABLE IF EXISTS testdb.%s", SingleStoreDialect.quoteIdentifier(records.iterator().next().topic())));
+        if (dropTableBeforeWrite) {
+            executeQuery(String.format("DROP TABLE IF EXISTS testdb.%s", SingleStoreDialect.quoteIdentifier(records.iterator().next().topic())));
+        }
         if (query != null) {
             executeQuery(query);
         }
